Extract enterprise payload limit module
